build: containerise all the things
We already had PostgreSQL and SQL Server containers which the tests ran against, which meant that having those databases installed locally on the host system was not required.

However, this still required that ODBC drivers, the driver manager and various versions of CPython and PyPy were installed locally for an adequate dev setup.

This commit is a bunch of QOL improvements to remove the need to install and configure drivers and a driver manager (although that is still possible if you really wanted to do so).

It also leans more into poetry and pyenv for the very small amount of stuff you might still want to do locally, e.g. pre-commit.